pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-06"
version = "0.1.0"
description = ""
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
faker = "^28.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
